apply refactoring suggestions
